moves Marker.svelte into guides/
