+++ src/app/patterns/body/elements/FirefoxReleaseVersionMarkers.svelte
@@ -1,6 +1,6 @@
 <script>
 import { getContext } from 'svelte';
-import Marker from '../../../../components/data-graphics/Marker.svelte';
+import Marker from '../../../../components/data-graphics/guides/Marker.svelte';
 import { firefoxVersionMarkers } from '../../../state/product-versions';
 
 import {
@@ -13,12 +13,14 @@ const xScale = getContext('xScale');
 
 let markers = [];
 firefoxVersionMarkers.subscribe((m) => { markers = m; });
+
 </script>
 
 
 <g class='firefox-release-version-markers'>
   {#if markers && markers.length}
-    {#each markers.map(({ date, label }) => ({ label, date: dateToBuildID(xScale, date) })).filter((d) => d.date !== undefined) as {label, date}, i (date)}
+    {#each markers.map(({ date, label }) => ({ label, date: dateToBuildID(xScale, date) }))
+      .filter((d) => d.date !== undefined) as {label, date}, i (date)}
       <Marker location={date}>
         {#if labels}
           {label}
